Fix conversion warnings for floats
Numbers should not be automatically converted to a less precise representation.
Rule of thumb: Use double for calculations and float for the output tree.

+++ src/Filters/ZJetCutsFilter.h
@@ -243,7 +243,7 @@ class BackToBackCut : public ZJetFilterBase
 	}
 
   private:
-	float backToBack;
+	double backToBack;
 };
 
 
@@ -281,22 +281,22 @@ class BetaCut : public ZJetFilterBase
 {
 	public:
 		virtual std::string GetFilterId() const override { return "BetaCut"; }
-		
+
 		BetaCut() : ZJetFilterBase(){};
-		
+
 		virtual void Init(ZJetSettings const& settings) override
 		{
 				ZJetFilterBase::Init(settings);
 				betaMax = settings.GetCutBetaMax();
 		}
-		
+
 		virtual bool DoesEventPass(ZJetEvent const& event, ZJetProduct const& product,
 							   ZJetSettings const& settings) const override
 		{
 			// Always true if there is no radiation jet in the event
 			return (product.GetRadiationJetCount(settings, event) > 0) ? (product.GetRadiationJet(settings, event, 0)->p4.Pt() < betaMax * product.m_z.p4.Pt()) : true;
 		}
-		
+
 	private:
 		float betaMax;
 };
